drbdsetup,v9,events2: start with UNKNOWN->xyz
So far "diffs" only contained a change "a->b" if the old state was know,
otherwise it was printed as "a".

It is easier to parse if we have a more regular structure, so if the old
state is not known, start with "UNKNOWN". This makes tokenizers a bit
easier.

It might be tempting to use a ternary operator for the color strings, it
is not. x ? "a", "b", "c" : "d", "e", "f" might not return what one
guesses.

While this introduces a new macro I think it is readable. The
alternative would be dedicated functions for every macro user as we
would need to evaluate different types.