drbdsetup,events2: fix --poll flag
09c5ee367513 drbdsetup,events2: keep track of tree of objects
and related commits changed the way that "drbdsetup events2" is
implemented. drbdsetup now remembers the objects it has seen and only
prints differences. This caused --poll to no longer output any objects
after the first dump of the state.

Fix this by clearing the remembered objects on each iteration of the
--poll loop.