drbdmeta: support activity logs with small final extents
Each activity log extent corresponds to a certain number of bitmap
bytes. The amount depends on the number of peers. For particular device
sizes and numbers of peers, the bitmap size may not be a multiple of the
number of bytes covered by each activity log extent. That is, there may
be a final extent which corresponds to fewer bitmap bytes.

For instance, consider an DRBD device of 18MiB net size, that is, the
size without metadata. Suppose the metadata has 7 bitmap peer slots.
This requires only one 4KiB bitmap block. This device is covered by 5
activity log extents. Each normal activity log extent corresponds to 896
bytes of bitmap. Hence the final activity log extent cannot correspond
to 896 bytes of bitmap, as shown:

|                            bitmap (4096 bytes)                          |
|   al-extent   |   al-extent   |   al-extent   |   al-extent   |al-extent|
|   896 bytes   |   896 bytes   |   896 bytes   |   896 bytes   |512 bytes|

In such a situation, when the final activity log extent was set,
drbdmeta failed with:
../shared/drbdmeta.c:1990:apply_al: ASSERT(bm_pos - bm_on_disk_pos <= chunk - extents_size) failed.

This occurred because it did not take the final truncated activity log
extent into account and hence would have written beyond the end of the
bitmap. The assert prevented this.

This issue has been present since:
528e1eafe27c drbdmeta apply-al: Support multiple peers

Prior to that, only one peer was supported. As a result, the activity
log extents always corresponded to 128 bytes of bitmap, and hence always
divided the bitmap size, since that is always a multiple of 4KiB.

This issue only occurs very rarely, as shown by the fact that it took 8
years to be discovered. It requires all of the following to hold:
* Number of bitmap peer slots not a power of 2
* Device size such that the final activity log extent is right at the
  end of the final bitmap block and overlaps the end of it
* Final activity log extent in use at the time DRBD is halted

Fix this by truncating the final activity log extent.