drbdmeta: fix unnecessary extra writes and reads when applying AL
This fixes an off-by-one error which caused a new chunk to be started
unnecessarily. This always occurred on the last iteration and could also
occur on previous iterations.

The issue was harmless, but fix it for correctness. Now the assert:
ASSERT(bm_pos - bm_on_disk_pos <= chunk - this_extent_size);
can be interpreted as expressing: "Ensure that the condition does not
hold which would cause us to need to start a new chunk." This is helpful
for understanding.