drbdsetup,v9: add support for events2 --poll without --now
There are now 4 distinct modes of invocation for events2:
* Without --poll or --now: Wait for updates from kernel and print them.
* --now: Print the current state and exit.
* --poll: Wait for updates from kernel or 'n' characters on stdin. Print
  updates when they are received. Fetch updated data when 'n' is read,
  but only show changed objects. This is useful with --statistics or
  --full because DRBD does not otherwise send updates when only the
  statistics change.
* --poll --now: Wait for 'n' characters on stdin. Print the full state
  each time one is read as a fresh set of "exists" events.

The --poll without --now combination is not supported on Windows.

The generic_get function is now implemented with generic_send and
generic_recv functions. These are used directly by the --poll
implementation to give it the necessary control over when to send new
initial state requests and when to wait for updates or input.

The implementation of --now has been tidied up. Instead of handling in
specially in print_event, we simply avoid subscribing to the multicast
group when it is set. We now filter received messages based on context
whenever cmd_id == DRBD_ADM_GET_INITIAL_STATE, because this command
returns unfiltered results, even when a context was provided.