WinDRBD: Make system root configurable via Registry
This patch substitutes references to path names (like
location of /etc for example) by a registry lookup on
the WinDRBD platform. For the Linux platform the path
names are also retrieved by C functions but these functions
return the original pathes as configured at compile time.

The Registry path is inside the WinDRBD services subkey
and the path is

	HKLM\System\CurrentControlSet\Services\WinDRBD\WinDRBDRoot

Note that the path must be given in cygwin notation so

	/cygdrive/c/windrbd instead of C:\WinDRBD

Also included in this commit are:
	.) When using block device interface (with a drive
	   letter in the devices clause) print a deprecation
	   warning. Block devices won't work with MSSQL server
	   that is why we will remove support from WinDRBD.
	   Users should use the disk interface instead.

	.) A configure check that checks for __free_fn_t
	   which does not exist on CygWin. We are not testing
	   for WinDRBD instead we are using a configure macro
           to check if the type is there (which makes it
           more portable).