drbdadm: Enforce maximum field with for strings when generating drbdsetup commands
In stacked setups, it generates concaternated hostnames of the lower hosts for the
peer connection name in the upper layer. That might exceed the 64 character limit.

It does not hurt to simply truncate that at 64 characters. Implement that in a
way that we would never output a too long string onto the drbdsetup command line.