drbd: fix segfault of wait-* family sub commands
Recent regression in 9.18.2 introduced by
  2880596f drbdadm: Enforce maximum field with for strings when generating drbdsetup commands

That changed "make_options" to take a "drbdsetup_ctx" for default options.
For the "wait-*" family of functions, we do not define that context, the NULL
is passed down into find_field(), which then in an attempt to iterate over
the default option list tries to dereference it.

fix: in find_field(), check options_def against NULL first.

We may consider to still add option contexts to the "wait-*" sub-command definitions.