drbdadm,v9,proxy: Allow multiple connections sharing a proxy node
Allow multiple connections to share a proxy node. The example is
three nodes: A, B, and C. Where A and B have a common proxy node
in site 1 and C has a dedicated proxy node in site 2.

So far this was refused by drbdadm with a very misleading error
message.

This works by setting the used_as_me flag on multple host
objects.