reactor/systemd: add "secondary-or-escalate"
If "drbdsetup secondary" fails,
a drbd-reactor based cluster would be blocked on this resource.

Optionally use either `FailureAction` directly (if you can use `ExecCondition`
for promotion) or `OnFailure=drbd-demote-or-escalate@%i.service` to
escalate to `reboot` (or `reboot-immediate`) if you need this.

For the paranoid setups, you'd need to have systemd use a WatchdogDevice,
set RuntimeWatchdogSec, RebootWatchdogSec, and possibly set JobTimeoutSec
and JobTimeoutAction on shutdown.target, reboot.target,
and the drbd-demote-or-escalate@.service.