build: fix spec file for RHEL 9.0
This fixes "bare words" errors like this:
error: bare words are no longer supported, please use "...":  systemd == "systemd"

It also fixes %systemd_postun, which needs an argument in the RHEL9.0
macros. Previously this had all kinds of forms, even %nil. All
distributions allow an arg, so set it to the service file.

Last but not least the build was broken because of C++ detection. If CXX
was set the detection did not work. Since forever, but RHEL9.0/rpmbuild
now sets CXX which triggeres that issue. The logic as is looks good
enough for me. Users can always "./configure --without-drbdmon ... &&
rpmbuild --without drbdmon ...".