drbd-demote-or-escalate@.service: ExitStopPost: adjust path for compatibility, ignore exit codes
I'd like to increase the chance that any log messages leading up to the
`FailureAction=reboot-immediate` will still be persisted, so the reason for the
reboot can clearly be determined.

A few manual iterations seem to suggest that using journalctl --sync as
ExecStopPost actually make a difference here, without it I frequently lost the
relevant logs.

Older systems seem to have it in /bin/journalctl (and do not search any PATH).
Newer systems symlink /usr/bin to /bin anyways.

Also ignore the exit code of the ExecStopPost.