reactor/systemd: add OnFailureJobMode and Conflicts
Avoid potential spurious reboots.

The OnFailure handler drbd-demote-or-escalate@.service
should Conflict with the failing service drbd-promote@.service
and should be scheduled as irreversible job,
so it won't "accidentally" be replaced by some other action/job.

Because that would likely cause things to "escalate" to the FailureAction.

You still need to enable it in some drop-in.