drbdmeta: initialize bitmap in chunks
Currently we (try to) initialize the bitmap area to all zeros
in one go using ioctl BLKZEROOUT.

With huge volumes (and high peer slot counts), the bitmap area
becomes quite large as well (several gigabytes).

Depending on the backend storage performance, that may take "some time".
Do this in chunks instead (I chose one gigabyte per ioctl for now),
so we can report progress, and can be interrupted between chunks.

If the BLKZEROOUT fails, we fall back to even slower direct-io pwrite()
in chunks of "buffer_size", which is 32 MiB at this time.