crm-fence-peer: fix timeout with Pacemaker 2.1: milli seconds vs seconds
crmadmin timeout was in milli seconds for <= 2.0.x,
but became a TIMESPEC with default seconds in >= 2.1.

Up to 2.0.4, atoi() was used, which effectively ignores "trailing garbage",
so we could get away with always appending "ms".
But with 2.0.5, it became g_option_context_parse G_OPTION_ARG_INT, which
"Cannot parse integer value “200ms” for --timeout" :-|

So grep the help message for "timeout.*milliseconds",
and if not present, append an explicit "ms" unit.

Also tolerate both ": ok" (2.1) and " (ok)" (older)
when matching the output string of crmadmin -S.