drbdadm: pass '--force' to some drbdmeta commands
Drbdadm already filters the options it passes to its backend commands.
It knows which options are accepted by a certain drbdsetup sub-command,
so it will not output an option that is not allowed for drbdsetup.

That allows us to add the '--force' to certain frontend command
definitions, so that it allows the flag on input.

Adding '--force' to
 * invalidate
 * wipe-md
 * apply-al
 * forget-peer