drbdsetup,drbdadm,v9: Allow drbdsetup secondary --force
Giving --force with the demote command causes any open IO requests
to terminate with error. Open operations are dissallowed until all
existing openers existed.

This change also introduces the 'fail-io' field to 'drbdsetup status'
and 'drbdsetup events2'

And introduces the resource option 'on-suspended-primary-outdated'
with the possible values 'disconnect' and 'force-secondary'.