drbdsetup,v9: fix susp_str()
susp_str() is supposed to deliver a textual representation of the suspend
reason. That worked fine as long as no one relied on the result of that
function for two calls.

Have a look at drbdsetup_events2.c:
			printf(" suspended:%s->%s",
					old_resource ? susp_str(&old_resource->info) : UNKNOWN_STRING,
					susp_str(&new_resource->info));

Since we support the "diff" mode it was broken. Fixing this by having all
the possible strings 16, always ready.