drbdadm,v9: fix sh-resource
I guess that command did not work in a long time, but an issue now
popped up because someone is actually using it in a script.

It actually "worked" as the structs are filled via `ctx_by_minor()`, but
a few lines below it failed in a check.

Fix by adding the flag that at least sounds like it would carry the
required semantics.

It should now work for res, res/vol, minor-num.

Alternative would have been to add an extra flag for that kind of
commands that strictly require a minor-num argument.