WinDRBD: DRBD meta data helpers for LINSTOR on Windows
There are 3 new commands:
	windrbd get-blockdevice-size prints size of a block device
	windrbd check-for-metadata returns exit code 0 if meta data is found
		1 on error and 2 if no meta data is found.
	windrbd wipe-metadata erases meta data (if found or -f is given)

All 3 accept drive letters (X:) and GUIDs (without the Volume{..} decoration)
as block devices (no CygWin /dev/sdX)

The reason for the helpers is:
	Getting size of a block device on Windows is not supported
	in Java. Also writing to a block device is not supported in
	Java, therefore we have this helpers.

	Also, getting size is very slow when done via PowerShell. When
	there are more than - let's say - 5 volumes performance
	is not acceptable.

LINSTOR then would require this commit (probably to be released with
WinDRBD 1.0.3 and up).

This commit does not touch Linux code.

Return values of check-for-metadata and wipe-metadata:

0 ... has meta data or cmd is wipe
1 ... error
2 ... has no meta data (check only)

Reason is that there are probably older windrbd versions installed
which return 1 if check-for-metadata is called (unknown command).

Always return 0 (success) if op is WIPE (except on error)

Reason is that LINSTOR calls wipe even if there is no meta data
which should not be considered an error.