drbdmeta: fix apply-al for bitmap sizes > 4GiB
A long time ago, the maximum supported bitmap size was 512 MiB.
Nowadays, we can do Petabyte volumes and multiple peer bitmaps.

Using "unsigned int" for the size of the bitmap is just wrong.

Use uint64_t for .bm_bytes and .bits_set members.
Use size_t for extent sizes, to not 32bit truncate multiplication results.

Should we ever want to support platforms with size_t != off_t,
and a bitmap that is not completely in-core,
we need to double check all size_t declarations.