parser,v9: deprecate named connections
While this was a nice idea I assume not too many users make use of it.
On the other hand I'm aware of code (and wrote such code myself) that
basically expects to find the peer's name in "conn-name" (the default
behavior).

Start deprecation by writing a warning (our err() is not the err() one
would expect, it is warn()) and already ignore the explicit name. This
is the middle ground between just writing a warning and already failing
hard in the parser. I'd like to keep it like this for at least one
non-bugfix release.