WinDRBD: Do not call set-mount-point any more.
Starting from WinDRBD 1.1.4 the drive letter ("mount point")
has to be assigned via the standard Windows tools. The driver
is building a virtual partition table around an existing
partition which makes Windows think it is a real SCSI disk
with a real partition table. This also causes Windows to
remember the drive letters.