WinDRBD: New command for terminating drbdsetup events2.
The new windrbd set-shutdown-flag command is meant to
be sent by the WinDRBD installer to signal that the
driver is about to be unloaded soon.

The driver then will fail all drbdsetup events2 and
further drbdadm / drbdsetup requests so that it can
upgrade the driver and the userland binaries.

It takes a flag argument so

	windrbd set-shutdown-flag 1

will stop drbdsetup events2 and

	windrbd set-shutdown-flag 0

will allow them again (in case unloading the driver failed)

Ioctl is for terminating drbdsetup events2 (and all other drbdsetup
commands) which is used to unload the WinDRBD driver (unloading fails
when there are still drbdsetup commands running).

Also print a warning message when driver is about to shut down.
So user isn't confused when drbdsetup events2 terminates with
an error.