drbdadm: drop use of GLOB_MAGCHAR, use strchr heuristic only
Fixup for
2022-09-05 4a1b5900 drbdadm: allow files from an expanded include glob to vanish

When using the `include` statement, if the glob did not match any file,
there is nothing to do, silently ignore.  Unless it was no glob, but a literal,
which we would expect to exist.

Also, there is a race between expanding a glob and accessing the file.
That also should not happen for literals, though.

Since we still had the heuristic anyways, because apparently |GLOB_MAGCHAR
does not happen for GLOB_NOMATCH returns, and there exist non-GNU libc that
don't (and likely won't) implement that extension, just forget about
(gl_flags & GLOB_MAGCHAR) but use the incomplete strchr heuristic only.