drbdadm: proxy-up: fix missing '-c' for plugins
the commit
bd0bac2b (drbdadm: Replace d_option list manipuations with STAILQ macros from queue.h, 2012-01-30)
accidentally dropped the "-c" in front of the "set plugin <connection-name> <counter> END".

No one noticed so far, as it would only be relevant on re-configuration,
which usually is done by proxy-down/proxy-up.

Still, this needs to be fixed so a "proxy-adjust",
once we have that, has at least a chance to work.