drbdadm: Fix validation of resync-after resource names
So far, drbdadm has tried to validate every resource named in a
resync-after option. The resync-after option might be present in the
host section for some peers. There it is entirely valid if there are
resources referenced which are not known locally. Fix that by not
looping over all volumes for all nodes but only for all volumes for
myself.

Since there is a complete check in _convert_after_option(), remove the
redundant code in validate_resource(). It gets inherited from there
to the volume option anyways.