drbd: Support user-defined block_size
Well, block storage devices have a particular sector size of block
size. The expression sector derives from "a part of a cylinder", from
when rotating rust dominated the world of block devices. Sector size
and block size stand here for the same concept.
This block size from the world of physical devices is sometimes called
hw_sector_size, PHY-SEC, physical block (sector) size, or even logical
block (sector) size. Usually, this block size is a fixed property of
the physical device.

Linux implements a cache on top of block devices, the "buffer cache".
The user can modify its block size using `blockdev --setbsz`. The help
text says "blocksize on file descriptor opening the block device".

DRBD needs to combine the physical block sizes of the backing
disks. (A diskless node does not contribute a physical block size on
its own.) In clusters with storage devices with different physical
block sizes, it is necessary to configure the maximal physical block
sizes on the DRBD level. Here is an example highlighting the need.

Let's say node A is diskless. It connects to node B, which has a
physical block size of 512 bytes. Then the user mounts the filesystem
on node A; the filesystem recognizes that it can do I/O in units of
512 bytes.
Later, node C joins the cluster with a physical block size of 4096
bytes. Now, suddenly DRBD starts to deliver I/O errors to the
filesystem if it chooses to do I/O on e.g., 512 or 1024 bytes.

The new configure option "block-size" comes to the rescue. You can
configure DRBD on node A to advertise a hw_sector_size of 4096 to the
filesystem before such a node becomes visible.