drbdadm_dump: Restructure dump_options() functions
So far we had dump_options() dump_options2() and the same for XML
format as well.

I simplify this for the call sites by replacing this with a single
function, which accepts a variable number of arguments.  When it gets
a subsection_name, then it produces a subsection for the second option
list and the following ones.

This is a preparation step for the next commit. There we need the
ability to combine options from 3 option lists into a single section
in the output.