drbdadm,v9: pass "--verbose" to "drbdsetup status"
Many users never use drbdsetup directly. The output of "drbdsetup status
--verbose" is nevertheless useful. So make it available via drbdadm too.

The "--verbose"/"-v" flag already has a meaning for drbdadm itself. Keep
that while also passing the flag down to drbdsetup.

This does cause a change of behavior, but users should not be relying on
the exact content and formatting of the "drbdadm status" output.