drbdmeta: drbdadm,v9: add repair-md command to fix up UUIDs
DRBD contained a bug until the following commit that leaves the metadata
in an unexpected state:
d8f5cab5b561 drbd: do not copy zero UUID from diskless peer when freeing peer slot

Add a command to drbdmeta for checking and repairing metadata that has
been corrupted in this way.

Also make this command accessible through drbdadm.

The flag --tentative is an alias to --dry-run. When either of these is
given, the metadata will not be changed. The name "tentative" is used to
allow the flag to be passed through drbdadm without conflicting with the
established meaning of --dry-run. This is similar to "drbdsetup connect
--tentative ..."