drbdsetup: represent meta-disk-index seperately
With the current implementation, the JSON output looks like this:

  "meta-disk": "internal"     (for internal metadata)
  "meta-disk": "/path/to/dev" (for external metadata)
  "/path/to/dev/": 3          (for external md with index, where 3 is the index)

This does not make much sense and is pretty difficult to correctly parse because
the "meta-disk" key just completely vanishes.

Instead, always output "meta-disk" and add "meta-disk-index" when applicable:

  "meta-disk": "/path/to/dev",
  "meta-disk-index": 3

This is technically a breaking change, but since it is basically
impossible to parse the previous format, we do not expect anyone to
actually use it.