Introduce default_types.h header
This serves as a centralized component for including basic definitions,
such as for some commonly used data types. It currently includes the
cstddef and cstdint headers and puts the size_t type and the fixed width
signed and unsigned integer types into the root namespace.
Many standard libraries and/or compilers did that implicitly, but that
changed at least partly in recent versions of GCC, causing compile
errors.