meta,v9: fix regression, allow attach after offline resize
The regression: drbd-utils versions 9.24 and 9.25 would remain "Diskless"
after an "incomplete resize", where before it would implicitly move the meta
data and attach.  Restore the expected behavior.

If the backend (LV) was resized, but the DRBD internal meta data
was not yet moved, the next attach would not find the internal meta data,
and would fail.

To cope with that, we remember the backend size in our "last known backing
device data" in /var/lib/drbd/*.lkbd, and check at the last known position.

If we find valid meta data at the last known position,
we can move it to the new end of the backing device.

That used to be implicitly done by the "apply-al" sub-command.

Since we introduced the "repair-md" sub-command, this implicit
check for offline resize and moving of the internal meta data
needs to take place there, because a "no valid meta data found" response
from "repair-md" would cancel the attach of that volume.