WinDRBD: Suspend/Resume I/O for a minor on the WinDRBD level (outside DRBD)
This adds two new commands to the windrbd utility.

When an application busy writes a block it may happen that
syncing never finished. In that case, run this suspend-io-for-minor,
wait for sync to finish and then run resume-io-for-minor
when sync has finished. Note that this is different from DRBD's
suspend-io command.