drbdadm,v9: fix unlikely NULL ptr deref
As a side project, I started to play with gcc-12's '-fanalyzer'
option. One of the findings here is a memory leak. As it is a
single-shot command that usually terminates quickly, we spent zero
energy in freeing allocated memory when we terminate the program due
to an error.

Moving forward, I will use '-fanalyzer -Wno-analyzer-malloc-leak' for
now.