new config option load-balance-paths
With this boolean setting, the user can enable a transport
implementation to establish all paths in parallel and distribute the
data packets over all available paths.

As of now, only the 'lb-tcp' transport module obeys this setting.

When the user does not select a transport module explicitly, DRBD
selects between tcp and 'lb-tcp' based on the 'load-balance-paths'
setting.

The user can explicitly configure 'transport "lb-tcp";' and
'load-balance-paths no;' which is then wire-protocol compatible with
the traditional 'tcp' transport.