adjust,v9: Improve adjust for diskless primaries
When drbd.conf changes how a diskless node reaches nodes holding
the data, applying the changes might be nontrivial.

Consider a diskless primary, having one peer with an UpToDate
disk. One might remove the connection between the diskless node and
that node with backing disk, and add another connection to another
diskful node at the same time. That is a valid change when the two
diskful nodes have a connection.

So far, `drbdadm adjust` failed to apply such a change since it
tried to remove the connection first, and add the other one after that.

Fixing that by building in the capability that `drbdadm adjust`
recovers from the first failure by retrying it after new connections
in the resource got added and established.