tests: sanitize environment
Tests did not execute successfully if for example the TZ was set to a
time zone that produces a '-' instead of an '+' in the output (e.g.,
users in the US). Fix that by setting LC_ALL and friends. UTC produces a
"+00:00" so we don't need to change the rest.