postparse: improve "file:line: error message" reporting
For some errors detected during the post-parse stage,
we reported the last line of the main config file,
when we meant to report the file and line
where the resource definition can be found.

Also reset the line, not only the config file name
when parsing the argument to --config-to-test.
Re-use "include_file()" for that instead of open-coding it.

We could potentially even re-use "include_stmt()" here
to allow globbing in --config-to-test.