drbdmeta: Dump and restor the members field
In a cluster with an even number of diskful nodes, a partition having
exactly half of the nodes with backing disks can keep quorum when it
had contact to the majority of the diskless nodes at the moment when
diskful node count dropped to 50%.

For DRBD to restore such a quorum after reboot, it got a new field in
the meta-data, the "members" field. It contains the node masks of the
visible nodes before the node restart.

This patch adds support for dumping and restoring the members field
from the meta-data.