drbdmeta: create-md: new options --effective-size and --diskful-peers
Various peers may have backing devices of different sizes.
That can be hard to avoid with LVM and different PE sizes on the nodes,
or when connecting nodes with different types of backends (zvol on some,
LVM on other nodes).

If we start to use a DRBD volume on one of the nodes with a larger
backing device, we may be unable to connect the smaller backends later,
because that then would "truncate" the size of the now already
initialized volumes on the nodes with the larger backing devices.

The "correct" way would be to first fully connect the DRBD cluster,
so DRBD can to its initial handshake and sync and "agree" on the
maximum size that can be supported by all nodes in the cluster.

But we (LINSTOR) wants to start using (or initializing with mkfs) a new
volume without waiting for all peers to be fully connected.  And Linstor
has enough context information to know what that size will be.

We can initialize the "last agreed to" respectively "effective" size
when creating the DRBD meta data. Additionally, we need to flag
our diskful peers as "MDF_PEER_DEVICE_SEEN".

That way, during attach, DRBD will know that there are some other
nodes with backing devices, and know that it does NOT know their current
size, but see the persisted "effective" size as "last agreed" size,
and settle for that until the time it learns about the actual backing
device size of the peers once connected.