events2: set may_promote:no promotion_score:0 while force-io-failure:yes
We had some reactor magic try to promote a node that was
"secondary --force"d before, and was still held open by old references.

Trying to promote in that state results in the error message
> State change failed: (-12) Device is held open by someone
In response to "drbdadm primary", that may be a bit "confusing" at first.

We should not report that state as "promotable".