drbdsetup,v9: show TLS in connection status
A user might want to make sure that TLS is actually used. With this commit
this information is displayed in the output of drbdsetup status.

Note that we generally do not show such information (net_conf stuff) in the
status output: we do not even show the addresses for the connection. For TLS
we make an exception: if you enable TLS it's probably important, and having a
simple way to confirm the configuration will help users in their setup.