DRBDmon: MDspTaskDetail: Reinitialize information if the selected task ID changes
If the MDspTaskDetails display remained in the display stack, but
another task ID was selected when the user returned to the display,
information about the previously selected task was shown, because
the information to be shown had already been prepared and was only
discarded when the display was closed.
This is not the intended behavior. Therefore, save the task ID
that the information has been prepared for, and reinitialize if
the selected task ID does not match the saved task ID.