drbdmeta: when asking pvs for pv_size, allow scan_lvs=1
Otherwise some resize / create-md / restore-md operations
would be refused (unless you --ignore-sanity-checks),
because drbdmeta is unavle to determine the "used size".

Use --readonly mode and allow to scan LVS so pvs -o pv_size
is willing to tell us the "used size" inside of a backend device,
in case that was used as PV in some nested fashion.