drbd: fix initialization of "bitmap area" for external meta data
Before, the zeroout bitmap paths would always use bm_bytes(), even for
"flexible external meta data", though in that case with the size of the
external meta data device instead of the backing device, which is not known at
this point.
We should instead initialize the "potential" bitmap area, which for
flexible external meta data is just "the rest of the device".

We have the cfg->bm_bytes member, use it!

cfg->bm_bytes should hold the on disk bitmap area size in bytes, rounded up to 4k.
That is the backing device size dependent bm_bytes() for "flexible internal",
the fixed size for (old and irrelevant now) fixed size internal
and "indexed" external, and just "the rest of the device" for "flexible external".