drbd_buildtag.h: only update if changed
It annnoyed my that drbdmon had to be partially rebuild and relinked for any
unrelated change, and even for no change at all, just because buildtag.h was
"touched", unconditionally, with every "make".

If there is no change, don't update it.

For buildtag.c, I don't do the same yet: it contains a build timestamp,
so would be changed in any case.

We could find the "most recent mtime" over all source files,
and use that as time stamp for "buildtag.c", but this has
not annoyed me enough yet to actually implement it.