configure.ac: add option for disabling keyutils
Some platforms (like Windows) do not have access to keyutils. Since it is only
used to make the configuration of TLS keys more convenient, we can swap the
implementation with a generic "numeric" one, as the DRBD API simply expects
a "key_serial_t", which is an alias for int32_t.

When not set, we try to detect the presence of keyutils and add support. If
explicitly requested, failure to find libkeyutils is an error.