expose drbd_info->dev_is_open in status and events2
A device may even be open read-only in Secondary role,
which would still cause a "down" to fail.
This was not recognizable from status or events2 output.

With the recent addition of "dev_is_open" to device_info,
we can now "know" whether a down or del-volume would fail
(because it is still in use) without trying.

Also update events2 test cases to expect the new "open:" text.