drbdsetup,v9: fix path output for json in show
so far if a connection had multiple paths, we generated multiple "path"
keys in the "connection" dictionary overwriting each other.

This fixes this by adding a "paths" list. Paths themselves are printed
as dictionary with "address", "port", and "family" keys. Both of these
fixes follow what we alredy have in "drbdsetup status --json".

As "drbdsetup show --json" is AFAIK mainly used in our internal tests I
think it is fair to introduce that breaking change. Non-json output
should be pixel perfect to what it was before.