fix out-of-bounds access in parsing
A call to sget_token advances the pointer to the to-be-parse text. It did check
for potential end-of-string situations, but only reported them if the token
could not be fully parsed. If the a full token was parsed right at the end of a
string, the implementation would advance the buffer pointer past the end of the
allocation, pointing past the NUL-byte.

A calling function could then try to get the next token which would lead to an
out-of-bounds access, that could potentially cause parsing of completely
unrelated data.

This issue was revealed when using -fsanitize=address.