drbdsetup,v9: ensure event consistency with peer devices
In the implementation of various wait-* commands we first fetch some initial
state, then listen to the events MC group for updates. However, we need to
ensure that they report a consistent state.

By first registering with the MC group, and only afterwards fetching the
initial state, we ensure that we do not miss any relevant events. Otherwise
there might be a small chance that certain events, such as removal of one of
to-be-waited-on entity is missed.