drbdsetup,v9: apply destroy events in wait-* commands
When waiting for a certain state, we need to apply both create and destroy
events to our remembered state, so that the evaluation of the condition has
an up-to-date view of the system.

This commit ensures we properly handle destroy events for peer devices when
waiting for a resource to be in sync. Otherwise the state might wait for peer
devices which are no longer configured, but still remembered in the internal
state.