drbdadm: Fix parsing of drbd-8.4 configs
The parser moves options to the drbd-9 style sections transparently
and renames them so they have new names in the 9 style. The
check_depricated_alias() function translates the version 8 option
names to version 9.

Missing was moving the 'resync' from the syncer section to the
peer_device_options, renaming it to 'resync-rate', and 'fencing' from
the disk section to the net section.

As the framework for these was already in place, adding these two is a
reasonably small change.