drbdmeta: check-resize with external meta data should not always "fail"
Various scripts, among them the DRBD OCF (pacemaker) resource agent,
call "drbdadm check-resize" just in case.

For external meta data, that would always result in an exit code 1,
because we do not have a .lkbd "last known block device" info file
for volumes with external meta data.

That can lead to spurious noise, for example the RA would report
four lines of "ERROR ... " because of the non-zero exit code,
while still continuing happily after that.

The behavior was introduced ten years ago in 2014,
when an intentionally ignored error condition (non-existence of that file)
was turned into an error return,
in response to some complaints of some static checker tool :-(