drbd-verify.py: A python script to automate verify of DRBD-replicas
This program automates running DRBD's verification operation on all
the resources of this host.

It is intended for interactive use, as it updates its progress display
every 3 seconds. It starts verifying operations between diskfull peers
by copying itself to those machines (/tmp/ folder) and running itself
there. For that, it requires logging into all peers by ssh without
providing a password.

It creates a file in the current working directory with the name
drbd-verify-result_YYYY-MM-DD_HHMM.json that contains the results in
JSON format.