drbdadm: add proxy options to 'add connection' command
Some options such as rcvbuf-size must be set before attempting to
connect. Proxy v3 used ugly hacks to allow time for the options to
arrive before connecting. Proxy v4 does not use this approach. Instead
it requires the options to be set on the command line for 'add
connection'.

The additional options are ignored by Proxy v3 with a "trailing garbage"
warning.

We continue to set the properties using 'set' commands as well to retain
compatibility with Proxy v3. Proxy v4 ignores 'set' commands which do
not change the value of the option.

So this is compatible with Proxy v3 and v4.