drbd.ocf: explicitly timeout crm_master IPC early
Some environments are notorious for timing out in crm_master
one out of ten thousand times. You don't want to know the details.

That would then cause a timeout on the monitor action,
and pacemaker would feel the need to "recover" from "resource failure",
spuriously restarting everything that depends on the DRBD resources.

If we fail to update the "master score" in time,
we can still report the operation result.

The next monitor action will happen,
and we get an other attempt at updating the master score.

If that update was relevant, worst case it is now delayed.
Better than the previous worst case, operation timeout
interpreted as resource failure.